refactor(capture): split Claude and Cursor bindings
