+++ crates/ctx-history-capture/src/provider/providers/claude/nativepath/source_backed/binding.rs
@@ -1,0 +1,28 @@
+use super::*;
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[serde(deny_unknown_fields)]
+pub(super) struct Binding {
+    pub(super) project_dir: PathBuf,
+    pub(super) key: ClaudeSessionKey,
+    pub(super) layout: SessionLayout,
+}
+
+pub(super) fn decode_binding(leaf: &JsonlFamilyLeaf) -> Result<Binding> {
+    let TypedKey::Bytes(bytes) = leaf.binding() else {
+        return Err(contract("Claude family binding is malformed"));
+    };
+    Ok(serde_json::from_slice(bytes)?)
+}
+
+pub(super) fn relative_to_authority(
+    authority: &ProviderSourceRoot,
+    path: &Path,
+) -> Result<PathBuf> {
+    path.strip_prefix(authority.named_path())
+        .map(Path::to_path_buf)
+        .map_err(|_| CaptureError::InvalidProviderTranscriptPath {
+            path: path.to_path_buf(),
+            reason: "Claude transcripts must remain below their selected authority",
+        })
+}
